Рефакторит MainPage
